% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_pvalue_right_tail}
\alias{bootstrap_pvalue_right_tail}
\title{Right-tailed bootstrap p-value with Davison & Hinkley (1997) (1 + B_ge)/(B+1).}
\usage{
bootstrap_pvalue_right_tail(boot_stats, stat)
}
\description{
Right-tailed bootstrap p-value with Davison & Hinkley (1997) (1 + B_ge)/(B+1).
}
\keyword{internal}
