% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fssd_test.R
\name{compute_fssd_unbiased_stat}
\alias{compute_fssd_unbiased_stat}
\title{Unbiased estimator of FSSD^2 (paper Eq. 2).}
\usage{
compute_fssd_unbiased_stat(tau_matrix)
}
\arguments{
\item{tau_matrix}{Numeric FSSD feature matrix.}
}
\value{
Numeric scalar.
}
\description{
Unbiased estimator of FSSD^2 (paper Eq. 2).
}
\examples{
tau <- matrix(c(1, 2, 3), ncol = 1)
compute_fssd_unbiased_stat(tau)
}
