% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fssd_test.R
\name{fssd_grid_search_scale}
\alias{fssd_grid_search_scale}
\title{Grid-search the squared kernel scale with locations fixed.}
\usage{
fssd_grid_search_scale(X, grads, V, kernel_obj_base, scale_grid, gamma)
}
\description{
Grid-search the squared kernel scale with locations fixed.
}
\keyword{internal}
