% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fssd_test.R
\name{fssd_init_locations_from_fit}
\alias{fssd_init_locations_from_fit}
\title{V ~ N(mu_hat, Sigma_hat + cov_reg * I) fitted to X.}
\usage{
fssd_init_locations_from_fit(X, J, cov_reg = 0)
}
\description{
V ~ N(mu_hat, Sigma_hat + cov_reg * I) fitted to X.
}
\keyword{internal}
