% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fssd_test.R
\name{fssd_kernel_with_scale}
\alias{fssd_kernel_with_scale}
\title{Re-instantiate a kernel with a new squared scale.}
\usage{
fssd_kernel_with_scale(base_kernel, scale2_k, has_kernel_param)
}
\description{
Re-instantiate a kernel with a new squared scale.
}
\keyword{internal}
