% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fssd_test.R
\name{fssd_tau_covariance}
\alias{fssd_tau_covariance}
\title{Paper plug-in covariance for FSSD null simulation.}
\usage{
fssd_tau_covariance(tau)
}
\description{
Paper plug-in covariance for FSSD null simulation.
}
\keyword{internal}
