% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stein_point_mcmc.R
\name{grwmetrop}
\alias{grwmetrop}
\title{Gaussian random-walk Metropolis.}
\usage{
grwmetrop(log_p, x0, S, m_iter)
}
\arguments{
\item{log_p}{Function returning log density values.}

\item{x0}{Initial state vector.}

\item{S}{Gaussian proposal covariance matrix.}

\item{m_iter}{Number of MCMC iterations.}
}
\value{
List containing chain states, log densities, acceptance indicators,
and evaluation counts.
}
\description{
Uses \code{S} as the Gaussian proposal covariance matrix.
}
\examples{
log_p <- function(X) -0.5 * rowSums(as.matrix(X)^2)
grwmetrop(log_p, x0 = 0, S = matrix(0.1), m_iter = 3)
}
