% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_classes.R
\name{kernel_generics}
\alias{kernel_generics}
\alias{eval_kernel}
\alias{grad_x_kernel}
\alias{trace_mixed_kernel}
\alias{cross_kernel}
\alias{grad_theta_v_kernel}
\title{Kernel building block generics}
\usage{
eval_kernel(obj, X, Y = NULL, ...)

grad_x_kernel(obj, X, Y = NULL, ...)

trace_mixed_kernel(obj, X, Y = NULL, ...)

cross_kernel(obj, X, grads, Y = NULL, grads_Y = NULL, ...)

grad_theta_v_kernel(obj, X, vj, grads_X, g_block, ...)
}
\arguments{
\item{obj}{A Stein kernel object.}

\item{X}{Numeric sample matrix.}

\item{Y}{Optional second sample matrix.}

\item{...}{Additional arguments passed to methods.}

\item{grads}{Score matrix for \code{X}.}

\item{grads_Y}{Optional score matrix for \code{Y}.}

\item{vj}{Feature location.}

\item{grads_X}{Score matrix for \code{X}.}

\item{g_block}{Gradient block used in FSSD optimisation.}
}
\description{
Low-level generic functions used to assemble Stein kernels and FSSD
objectives.
}
\examples{
X <- matrix(c(-1, 0, 1), ncol = 1)
grads <- -X
kernel <- stein_kernel(type = "gaussian_rbf", h = 1)
eval_kernel(kernel, X)
grad_x_kernel(kernel, X)
trace_mixed_kernel(kernel, X)
cross_kernel(kernel, X, grads)
grad_theta_v_kernel(kernel, X, vj = 0, grads_X = grads, g_block = grads)
}
