% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stein_helpers.R
\name{max_offdiag_from_kernel_matrix}
\alias{max_offdiag_from_kernel_matrix}
\title{Max off-diagonal value from a kernel matrix}
\usage{
max_offdiag_from_kernel_matrix(k_mat)
}
\description{
Max off-diagonal value from a kernel matrix
}
\keyword{internal}
