% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgd.R
\name{print.svgd}
\alias{print.svgd}
\title{Print an SVGD fit}
\usage{
\method{print}{svgd}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=svgd]{svgd()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an SVGD fit
}
