% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{resolve_markov_change_prob}
\alias{resolve_markov_change_prob}
\title{Resolve Markov sign-flip probability.}
\usage{
resolve_markov_change_prob(change_prob)
}
\description{
Resolve Markov sign-flip probability.
}
\keyword{internal}
