% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_model.R
\name{row_logsumexp}
\alias{row_logsumexp}
\title{Row-wise log-sum-exp.}
\usage{
row_logsumexp(log_mat)
}
\arguments{
\item{log_mat}{Numeric matrix of log values.}
}
\value{
Numeric vector of row-wise log-sum-exp values.
}
\description{
Row-wise log-sum-exp.
}
\examples{
row_logsumexp(matrix(c(0, 1, 2, 3), nrow = 2))
}
