% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stein_points.R
\name{sample_proposal_box}
\alias{sample_proposal_box}
\title{Draw samples from the optimiser proposal inside \verb{[lb, ub]}.}
\usage{
sample_proposal_box(
  n,
  lb,
  ub,
  mu0,
  Sigma0,
  sigsq,
  X_curr,
  delay,
  max_oversample = 200L
)
}
\arguments{
\item{n}{Number of samples to draw.}

\item{lb, ub}{Lower and upper bounds.}

\item{mu0, Sigma0}{Initial Gaussian proposal mean and covariance.}

\item{sigsq}{Local proposal variance.}

\item{X_curr}{Currently selected points.}

\item{delay}{Number of selected points before using local proposals.}

\item{max_oversample}{Maximum number of proposal batches.}
}
\value{
Numeric matrix of proposed samples.
}
\description{
Draw samples from the optimiser proposal inside \verb{[lb, ub]}.
}
\examples{
sample_proposal_box(
  n = 3, lb = -1, ub = 1, mu0 = 0, Sigma0 = matrix(0.1),
  sigsq = 0.1, X_curr = matrix(0, ncol = 1), delay = 2
)
}
