% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fssd_test.R
\name{compute_tau}
\alias{compute_tau}
\title{Compute the FSSD feature matrix tau (n x dJ).}
\usage{
compute_tau(X, grads, V, kernel_obj)
}
\arguments{
\item{X}{Numeric sample matrix.}

\item{grads}{Score matrix for \code{X}.}

\item{V}{Matrix of FSSD test locations.}

\item{kernel_obj}{Stein kernel object.}
}
\value{
Numeric FSSD feature matrix.
}
\description{
Row i is \code{tau(x_i) = vec(Xi(x_i))} with
\verb{[Xi(x_i)]_\{l,j\} = xi_\{p,l\}(x_i, v_j) / sqrt(dJ)}, where
\verb{xi_\{p,l\}(x, v) = score_l(x) * k(x, v) + d/dx_l k(x, v)}.
}
\examples{
X <- matrix(c(-1, 0, 1), ncol = 1)
grads <- -X
V <- matrix(0, ncol = 1)
kernel <- stein_kernel(type = "gaussian_rbf", h = 1)
compute_tau(X, grads, V, kernel)
}
