% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_classes.R
\name{custom_stein_kernel}
\alias{custom_stein_kernel}
\title{Construct a custom Stein kernel}
\usage{
custom_stein_kernel(
  eval_fn,
  grad_x_fn,
  trace_mixed_fn,
  grad_theta_v_fn = NULL,
  custom_grad_mode = c("error", "analytic", "numeric")
)
}
\arguments{
\item{eval_fn}{Function evaluating the base kernel.}

\item{grad_x_fn}{Function evaluating the gradient with respect to \code{X}.}

\item{trace_mixed_fn}{Function evaluating the mixed derivative trace.}

\item{grad_theta_v_fn}{Optional gradient for FSSD feature optimisation.}

\item{custom_grad_mode}{How custom gradients are handled.}
}
\value{
A custom Stein kernel object.
}
\description{
Construct a custom Stein kernel
}
\examples{
eval_fn <- function(X, Y = NULL, precon = NULL) {
  X <- as.matrix(X)
  Y <- if (is.null(Y)) X else as.matrix(Y)
  diff <- outer(X[, 1], Y[, 1], "-")
  exp(-0.5 * diff^2)
}
grad_fn <- function(X, Y = NULL, precon = NULL) {
  X <- as.matrix(X)
  Y <- if (is.null(Y)) X else as.matrix(Y)
  diff <- outer(X[, 1], Y[, 1], "-")
  array(-diff * exp(-0.5 * diff^2), dim = c(nrow(X), nrow(Y), 1))
}
trace_fn <- function(X, Y = NULL, precon = NULL) {
  X <- as.matrix(X)
  Y <- if (is.null(Y)) X else as.matrix(Y)
  diff <- outer(X[, 1], Y[, 1], "-")
  (1 - diff^2) * exp(-0.5 * diff^2)
}
custom_stein_kernel(eval_fn, grad_fn, trace_fn)
}
