% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stein_points.R
\name{fmin_grid}
\alias{fmin_grid}
\title{Grid-search optimiser.}
\usage{
fmin_grid(lb, ub, n0 = 100, grow = TRUE)
}
\arguments{
\item{lb, ub}{Lower and upper bounds for candidate points.}

\item{n0}{Initial grid size per dimension.}

\item{grow}{Logical; whether to increase grid size as points are selected.}
}
\value{
Optimizer function.
}
\description{
Evaluates the objective on a Cartesian grid in \verb{[lb, ub]}. With \code{grow = TRUE},
the grid resolution increases slowly as more points are selected.
}
\examples{
fmin_grid(lb = -1, ub = 1, n0 = 5, grow = FALSE)
}
