% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksd_v_test.R
\name{ksd_v_statistic}
\alias{ksd_v_statistic}
\title{KSD V-statistic}
\usage{
ksd_v_statistic(U_mat)
}
\arguments{
\item{U_mat}{Numeric square matrix with entries \code{u_q(x_i, x_j)}.}
}
\value{
Numeric scalar \code{n * mean(U_mat)}.
}
\description{
KSD V-statistic
}
\examples{
U <- matrix(c(1, 0.2, 0.3, 0.2, 1, 0.4, 0.3, 0.4, 1), 3, 3)
ksd_v_statistic(U)
}
