% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_model.R
\name{posteriorgmm}
\alias{posteriorgmm}
\title{Calculate posterior component probabilities for a GMM}
\usage{
posteriorgmm(model = NULL, X = NULL)
}
\arguments{
\item{model}{Gaussian mixture model created by \code{gmm()}.}

\item{X}{Numeric vector or matrix of samples.}
}
\value{
Matrix of posterior component probabilities.
}
\description{
Calculate posterior component probabilities for a GMM
}
\examples{
model <- gmm()
X <- rgmm(model, n = 5)
posteriorgmm(model, X)
}
