% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_classes.R
\name{print.SteinKernel}
\alias{print.SteinKernel}
\title{Print a Stein kernel}
\usage{
\method{print}{SteinKernel}(x, ...)
}
\arguments{
\item{x}{A \code{SteinKernel} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Stein kernel
}
\examples{
stein_kernel("imq", c = 1, beta = -0.5)
}
