% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stein_points.R
\name{stein_kernel_imq_score}
\alias{stein_kernel_imq_score}
\title{k3: IMQ-score Stein kernel (alpha + ||score_p(x) - score_p(y)||^2)^beta.}
\usage{
stein_kernel_imq_score(alpha = 1, beta = -0.5, hess_log_p)
}
\arguments{
\item{alpha}{Positive offset parameter.}

\item{beta}{Exponent in \verb{(-1, 0)}.}

\item{hess_log_p}{Hessian function for the log density.}
}
\value{
A Stein kernel object.
}
\description{
\code{hess_log_p(X)} must return an n x d x d array of Hess(log p)(x_i).
}
\examples{
hess_log_p <- function(X) array(-1, dim = c(nrow(as.matrix(X)), 1, 1))
stein_kernel_imq_score(alpha = 1, beta = -0.5, hess_log_p = hess_log_p)
}
