% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stein_points.R
\name{stein_kernel_inverse_log}
\alias{stein_kernel_inverse_log}
\title{k2: inverse-log Stein kernel (alpha + log(1 + ||x - y||^2))^beta.}
\usage{
stein_kernel_inverse_log(alpha = 1, beta = -1)
}
\arguments{
\item{alpha}{Positive offset parameter.}

\item{beta}{Negative exponent.}
}
\value{
A Stein kernel object.
}
\description{
k2: inverse-log Stein kernel (alpha + log(1 + ||x - y||^2))^beta.
}
\examples{
stein_kernel_inverse_log(alpha = 1, beta = -1)
}
