% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_classes.R
\name{stein_kernel_matrix}
\alias{stein_kernel_matrix}
\title{Compute a Stein kernel matrix}
\usage{
stein_kernel_matrix(kernel, X, grads, Y = NULL, grads_Y = NULL, ...)
}
\arguments{
\item{kernel}{A Stein kernel object.}

\item{X}{Numeric sample matrix.}

\item{grads}{Score matrix for \code{X}.}

\item{Y}{Optional second sample matrix.}

\item{grads_Y}{Optional score matrix for \code{Y}.}

\item{...}{Additional arguments passed to kernel methods.}
}
\value{
Numeric matrix of Stein kernel values.
}
\description{
Compute a Stein kernel matrix
}
\examples{
X <- matrix(c(-1, 0, 1), ncol = 1)
grads <- -X
kernel <- stein_kernel(type = "gaussian_rbf", h = 1)
stein_kernel_matrix(kernel, X, grads)
}
