% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steinsampling-package.R
\docType{package}
\name{steinsampling-package}
\alias{steinsampling}
\alias{steinsampling-package}
\title{steinsampling package}
\description{
Kernel Stein discrepancy goodness-of-fit tests and Stein-method-based
sampling methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/junhao7622/Stein_Sampling}
  \item Report bugs at \url{https://github.com/junhao7622/Stein_Sampling/issues}
}

}
\author{
\strong{Maintainer}: Junhao Gao \email{jug049@ucsd.edu}

Authors:
\itemize{
  \item Junhao Gao \email{jug049@ucsd.edu}
  \item Ery Arias-Castro \email{eariascastro@ucsd.edu}
}

}
\keyword{internal}
